[alsa-smartx-plugin] Use always increasing hw pointer internally that wraps around at ALSA boundary value instead of buffer level.
This allows to get rid of the patch to the alsa-lib, which introduced a new function pointer for the io-plug mechanism. The patch
in the alsa-lib was used in the past to get rid of the issue that the io-plug could not distinguish between a full and an empty
buffer. The alsa-lib version 1.1.6 is required with this version of the alsa-smartx-plugin to work correctly as v1.1.6 contains
some very important patches applied to the io-plug implementation in the alsa-lib.