[alsa-smartx-plugin] Don't reset hardware pointer when initializing application pointer
Capture devices query mHwPtrWrite in order to determine how many
frames are available. Instead of initializing this value to 0
when setting software parameters, we should set it to mAvailMin so that
capture devices don't hang waiting for the buffer to be filled. Ideally
we would not modify the hardware pointer at all when initializing the
application but have a separate application pointer that is subtracted
from the hardware pointer to determine the available frames.